Moved GrassColor and FoliageColor to client/renderer/
