+++ source/client/renderer/FoliageColor.cpp
