+++ source/world/tile/LeafTile.hpp
@@ -27,5 +27,4 @@ public:
 
 	int* field_70;
 	int field_74;
-	bool m_bUseBiomeColors;
 };
